Refactor the getNum function
The displayInput for e.target.value was still running even when it is a key button that was pressed because the e.target.value still has referenced value making it true for the conditional statement. This error has been fixed by dividing the function into two, one for e.target.value and one for e.key.